// src/models/IoK8sApimachineryPkgApisMetaV1LabelSelector/types.ts
import { IoK8sApimachineryPkgApisMetaV1LabelSelectorRequirementModelOptions as IoK8sApimachineryPkgApisMetaV1LabelSelectorRequirement } from '../IoK8sApimachineryPkgApisMetaV1LabelSelectorRequirement';

export interface IoK8sApimachineryPkgApisMetaV1LabelSelectorModelOptions {
  matchExpressions?: Array<IoK8sApimachineryPkgApisMetaV1LabelSelectorRequirement>;
  matchLabels?: { [key: string]: string };
}